Move policies to separate module
